Added test case for config file without a new line at the end
Added test case for configuration file without a new line at the end. This case was covered in JavasScript but not Java.
This requirement is stated in the program description repo as:
"All the lines must have the same length and must be terminated by [\n] (INCLUDING THE LAST ONE!)"